% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{apply_mask}
\alias{apply_mask}
\title{Aplicar configuración a datos}
\usage{
apply_mask(data, cfg)
}
\arguments{
\item{data}{Objeto \code{sf} POINT.}

\item{cfg}{Objeto \code{geomask_config}.}
}
\description{
Aplicar configuración a datos
}
