% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateGMpoints.R
\name{calculateGMpoints}
\alias{calculateGMpoints}
\title{Calculate geomasked points}
\usage{
calculateGMpoints(myshps, maskvars)
}
\arguments{
\item{myshps}{List of shapefiles}

\item{maskvars}{List of settings for calculating masked locations.}
}
\description{
Calculate geomasked points
}
