% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomask-package.R
\docType{package}
\name{geomask-package}
\alias{geomask}
\alias{geomask-package}
\title{geomask: Spatial Geomasking and Anonymization Tools}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Herramientas unificadas para aplicar, evaluar y documentar estrategias de geomasking (enmascaramiento espacial) en datos georreferenciados sensibles. Compatible con objetos 'sf' y pensado para investigación en salud pública, epidemiología y ciencias sociales. Incluye métodos clásicos (donut, gaussiano, intercambio contextual), enmascaramiento informado por densidad, agregación adaptativa, anonimización basada en grafos y métricas de privacidad-utilidad.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JDConejeros/geomask}
  \item \url{https://jdconejeros.github.io/geomask/}
  \item Report bugs at \url{https://github.com/JDConejeros/geomask/issues}
}

}
\author{
\strong{Maintainer}: José Daniel Conejeros \email{jo.conejeros@gmail.com} (\href{https://orcid.org/0000-0003-3402-4361}{ORCID})

Authors:
\itemize{
  \item José Daniel Conejeros \email{jo.conejeros@gmail.com} (\href{https://orcid.org/0000-0003-3402-4361}{ORCID})
}

}
\keyword{internal}
