% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomask.R
\name{geomask}
\alias{geomask}
\title{The NYS Geomasking Tool (geomasker)}
\description{
A method for shifting point-level data to mask locations using user-defined
parameters with minimal knowledge of or experience using R.
}
\details{
Geomasking goals:
\itemize{
\item simplify and standardize the process
\item be in a format accessible to anyone
\item not require prior experience with R
\item be customizable for advanced R users
}

The geomasker was made possible through funding made available to the New
York State Department of Health / HRI by the CDC through
grant # NUE1EH001341-03-01.
Its content is solely the responsibility of the authors and does not
necessarily represent the official views of the CDC.
}
\author{
Abigail Stamm, Bureau of Environmental and Occupational
Epidemiology, New York State Department of Health \cr
Gwen Babcock, Bureau of Environmental and Occupational
Epidemiology, New York State Department of Health \cr
}
