% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomask_boundary}
\alias{geomask_boundary}
\title{Límites comunales simplificados (RM)}
\format{
Un objeto \code{sf} POLYGON con columnas \code{zone_id}, \code{comuna}, \code{pop}.
}
\usage{
geomask_boundary
}
\description{
Polígonos construidos como envolvente convexa por comuna a partir de
establecimientos de salud públicos (referencia espacial, no límites oficiales).
}
\keyword{datasets}
