% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{geomask_config}
\alias{geomask_config}
\title{Configuración reproducible de geomasking}
\usage{
geomask_config(method = "donut", params = list(), seed = NULL, boundary = NULL)
}
\arguments{
\item{method}{Nombre del método (\code{"donut"}, \code{"gaussian"}, etc.).}

\item{params}{Lista de parámetros del método.}

\item{seed}{Semilla.}

\item{boundary}{Ruta o objeto sf para límite (opcional).}
}
\value{
Objeto S3 \code{geomask_config}.
}
\description{
Configuración reproducible de geomasking
}
