% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomask_pop}
\alias{geomask_pop}
\title{Grilla de densidad poblacional sintética (RM)}
\format{
Objeto \code{sf} POLYGON con columnas \code{cell_id}, \code{density}.
}
\usage{
geomask_pop
}
\description{
Grilla de densidad poblacional sintética (RM)
}
\keyword{datasets}
