% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geomask_services}
\alias{geomask_services}
\title{Establecimientos de salud de ejemplo (RM, submuestra)}
\format{
Un objeto \code{sf} POINT con metadatos mínimos para tutoriales.
}
\usage{
geomask_services
}
\description{
Establecimientos de salud de ejemplo (RM, submuestra)
}
\keyword{datasets}
