% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomasking.R
\name{geomasking}
\alias{geomasking}
\title{Apply Gaussian or Uniform geomasking to a set of spatial coordinates.}
\usage{
geomasking(locations, displacement, delta)
}
\arguments{
\item{locations}{A two column matrix containing the locations (coordinates x
and y) to which apply geomasking.}

\item{displacement}{The type of geomasking to be applied: either "gaussian"
or "uniform".}

\item{delta}{A number that specify the standard deviation of the positional
error in the case of Gaussian geomasking or the maximum displacement
distance in the case of Uniform geomasking.}
}
\value{
A matrix with the displaced locations.
}
\description{
Apply Gaussian or Uniform geomasking to a set of spatial coordinates.
}
