% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomasking.R
\name{mask_gaussian}
\alias{mask_gaussian}
\title{Geomasking gaussiano (perturbación normal)}
\usage{
mask_gaussian(data, sigma, seed = NULL)
}
\arguments{
\item{data}{Objeto \code{sf} POINT.}

\item{sigma}{Desviación estándar del desplazamiento en x e y (metros).}

\item{seed}{Semilla opcional.}
}
\value{
Objeto \code{sf} enmascarado.
}
\description{
Geomasking gaussiano (perturbación normal)
}
\examples{
data(geomask_pts, package = "geomask")
masked <- mask_gaussian(geomask_pts, sigma = 400, seed = 3)
}
