% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomasking.R
\name{mask_location_swapping}
\alias{mask_location_swapping}
\title{Intercambio de ubicaciones dentro de grupos contextuales}
\usage{
mask_location_swapping(data, context_vars, min_dist = 0, seed = NULL)
}
\arguments{
\item{data}{Objeto \code{sf} POINT.}

\item{context_vars}{Variables que definen grupos de permutación.}

\item{min_dist}{Distancia mínima entre ubicación original y asignada (m).}

\item{seed}{Semilla opcional.}
}
\value{
Objeto \code{sf} con geometrías permutadas.
}
\description{
Intercambio de ubicaciones dentro de grupos contextuales
}
