% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{plot_displacement_distribution}
\alias{plot_displacement_distribution}
\title{Distribución de desplazamientos}
\usage{
plot_displacement_distribution(masked)
}
\arguments{
\item{masked}{Objeto enmascarado con \code{geomask_displacement_m}.}
}
\description{
Distribución de desplazamientos
}
