% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGMcompare.R
\name{plotGMcompare}
\alias{plotGMcompare}
\title{Draw the map comparing original and masked point locations}
\usage{
plotGMcompare(
  bound,
  original = NULL,
  shifted = NULL,
  maskvars,
  closemap = FALSE
)
}
\arguments{
\item{bound}{Boundary layer.}

\item{original}{Original points or areas spatial layer.}

\item{shifted}{Shifted points or areas spatial layer.}

\item{maskvars}{List of objects needed to calculate masked points. It
contains the numbers min and max, the boolean projection,
and the strings unit, point_id, and bound_id.}

\item{closemap}{Boolean denoting whether to close the map window after
the map is drawn and saved.}
}
\value{
Recorded plot object.
}
\description{
This function draws two maps on top of each other, so that point locations
can be compared. It is designed to be saved to a recordPlot object for
later rendering in a PDF rather than displayed directly.
}
\details{
If plotting directly, the display may be distorted. For best results, save
to a recordPlot object and write to PDF.
}
\examples{

maskvars <- list(min = 100, max = 1000, unit = "meters",
                 point_id = "POINTID", bound_id = "GEOID10")
ot <- tigris::tracts("NY", "Onondaga", year = 2010)
ol <- tigris::landmarks("NY", "point", year = 2015)
t <- unlist(sf::st_contains(ot, ol))
ol <- ol[sample(t, size = 20), ]
myshps <- list(point = ol, bound = ot)
myshps <- calculateGMpoints(myshps = myshps, maskvars = maskvars)

plotGMcompare(bound = myshps$bound,
              original = myshps$old_full,
              shifted = myshps$new_full,
              maskvars = maskvars)

}
