% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmci.R
\name{qmci}
\alias{qmci}
\title{Likelihood calculation with Quasi Monter Carlo Integration}
\usage{
qmci(data, sequence, sigma2, phi, nugget, mu, delta, kappa)
}
\description{
This function calculates the liklelihood of the geostatistical model in
presence of positional error using Quasi Monte Carlo Integration. It is used
internally in the optimisation routine.
}
