% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGMprogram.R
\name{runGMprogram}
\alias{runGMprogram}
\title{Run the geomasker}
\usage{
runGMprogram(bgcol = "thistle1", buttoncol = "plum2", settings = NULL)
}
\arguments{
\item{bgcol}{Text string containing UI background color.}

\item{buttoncol}{Text string containing UI button color.}

\item{settings}{The filepath to an Rdata file previously created by the
Geomasker.}
}
\description{
This function runs the full geomasking tool.
}
\details{
This function can read in an optional settings file from a prior run if,
for example, you would like to run the geomasker with two different
distances, but keep all other settings the same.
}
\examples{

if (interactive()) {
# this code will run the full program
runGMprogram()
}

}
