% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectGMdistances.R
\name{selectGMdistances}
\alias{selectGMdistances}
\title{Select geoMasking distances}
\usage{
selectGMdistances(
  step = 3,
  min = "1,000",
  max = "10,000",
  unit = "meters",
  backopt = TRUE,
  bgcol = "lightskyblue3",
  quitopt = "Quit",
  buttoncol = "cornflowerblue"
)
}
\arguments{
\item{step}{Step in the program.}

\item{min}{Number denoting minimum masking distance.}

\item{max}{Number denoting maximum masking distance.}

\item{unit}{String denoting unit of distance.}

\item{backopt}{Boolean denoting whether to include the back button.}

\item{bgcol}{Text string containing UI background color.}

\item{quitopt}{Text string for the cancel button.}

\item{buttoncol}{Text string containing UI button color.}
}
\description{
This function is modified from a similar function in gatpkg. It creates
a user dialog to read the minimum and maximum distances, as well as the
distance units, for masking.
}
\examples{

if (interactive()) {
selectGMdistances()
}

}
