% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_spdata.R
\name{sim_spdata}
\alias{sim_spdata}
\title{Simulation of gaussian spatial data using Cholesky decomposition}
\usage{
sim_spdata(locations, cov_model = "matern", cov_pars = c(1, 0.16),
  nugget = 0.5, kappa = 0.5)
}
\arguments{
\item{locations}{A two column matrix containing the locations (coordinates x
and y).}

\item{cov_model}{A string indicating the type of the correlation function.
For the available choices see \code{\link[geoR]{cov.spatial}}.}

\item{cov_pars}{A numeric vector with 2 elements with the covariance
parameters. The first element corresponds to the variance parameter
 \eqn{\sigma^2}. The second element or corresponds to the range parameter
  \eqn{\phi} of the correlation function.}

\item{nugget}{Value of the nugget parameter \eqn{\tau^2}.}

\item{kappa}{Numerical value for the additional smoothness parameter of
the correlation function. Only required by the following correlation
functions: "matern", "powered.exponential", "cauchy", "gencauchy" and
"gneiting.matern".}
}
\value{
A numeric vector containing the simualted spatial data.
}
\description{
Given a set of spatial coordinates this function generates gaussian distributed data whose spatial structure is defined by the user through the specification of a spatial correlation function.
}
