% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splm_sim_aggr.R
\name{splm_sim_aggr}
\alias{splm_sim_aggr}
\title{Spatial model for spatially aggregated data}
\usage{
splm_sim_aggr(data, mc_points = 100, tau.sq = 0, ncovariates = 2,
  beta = c(3.5, -1.2), aggr.cell = 9, fix.nugget = F, fix.nug = 0.5,
  ini = c(0.5, 0.5), message = F)
}
\arguments{
\item{data}{A numeric vector of spatial data.}

\item{mc_points}{Number of Monte Carlo points to calculate the correlation matrix.}

\item{tau.sq}{Numeric value for the nugger parameter \eqn{\nugget^2}.}

\item{ncovariates}{Number of covariates to simulate.}

\item{beta}{Values for the beta parameters.}

\item{aggr.cell}{To how many cells the points should be aggregated.}

\item{fix.nugget}{A logical value. Should the nugget be fixed?}

\item{fix.nug}{If fix.nugget = TRUE then a numeric value should be provided.}

\item{ini}{Initial parameters values for the optimisation algorithm.}

\item{message}{If FALSE suppress all the messages from the fitting algorithm.}
}
\description{
This function provide  model based parameters estiamtion for a geostatical model when spatial data are avialable at a corarser scale than their natural resolution. It is mainly for simulation purposing and testing.
}
