% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_metrics.R
\name{utility_summary}
\alias{utility_summary}
\title{Resumen de utilidad (envolvente)}
\usage{
utility_summary(original, masked, ...)
}
\arguments{
\item{original, masked}{Objetos \code{sf} POINT.}

\item{...}{Argumentos adicionales para sub-funciones.}
}
\value{
Objeto S3 \code{geomask_utility}.
}
\description{
Resumen de utilidad (envolvente)
}
