% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varioadj.R
\name{varioadj}
\alias{varioadj}
\title{Variogram fit adjusting for positional error}
\usage{
varioadj(data, locations, displacement = "gaussian", delta, kappa, ini, ...)
}
\arguments{
\item{data}{A numeric vector of spatial data.}

\item{locations}{A two column matrix containing the locations (coordinates x
and y).}

\item{displacement}{The type of geomasking to be applied: either "gaussian"
or "uniform".}

\item{delta}{A number that specify the standard deviation of the positional
error in the case of Gaussian geomasking or the maximum displacement
distance in the case of Uniform geomasking.}

\item{kappa}{Numerical value for the additional smoothness parameter of the
matern correlation function.}

\item{ini}{Initial values for the parameters to be passed to the optimisation
algorithm.}

\item{...}{Control argumenets for the optimiser.}
}
\value{
A numeric vector containg the estimated parameters.
}
\description{
This function fit the parameters of a variogram through N-weighetd least
squares taking also into account the presence of positional error.
}
