% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{write_geomask_config}
\alias{write_geomask_config}
\title{Escribir configuración a YAML}
\usage{
write_geomask_config(cfg, path)
}
\arguments{
\item{cfg}{Objeto \code{geomask_config}.}

\item{path}{Ruta de salida.}
}
\description{
Escribir configuración a YAML
}
