% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirmGMsettings.R
\name{confirmGMsettings}
\alias{confirmGMsettings}
\title{Confirm geomasking settings}
\usage{
confirmGMsettings(
  maskvars,
  filevars,
  savekml = FALSE,
  step = 0,
  bgcol = "lightskyblue3",
  quitopt = "Quit",
  buttoncol = "cornflowerblue"
)
}
\arguments{
\item{maskvars}{List of objects needed to calculate masked points. It
contains the numbers min and max, the boolean projection,
and the strings unit, point_id, and bound_id.}

\item{filevars}{List of strings, including the names and paths of
shapefiles to read in and save to, without extensions.}

\item{savekml}{Boolean denoting whether or not to save a KML file.}

\item{step}{Integer step in the GAT program, for help reference.}

\item{bgcol}{Text string containing UI background color.}

\item{quitopt}{Text string for the cancel button.}

\item{buttoncol}{Text string containing UI button color.}
}
\description{
This function opens a dialog window for the user to select whether the
selected settings for GAT are correct.
}
\details{
This function reads in the lists created during the user input portion of
the GAT tool. As such, it requires that the inputted lists contain the same
elements as those generated by the Geomasker.
}
\examples{

if (interactive()) {
maskvars <- list(min = 100, max = 1000, unit = "meters",
                 point_id = "POINTID", bound_id = "GEOID10")
filevars <- list(
  pointin = paste(getwd(), "points", sep = "/"),
  boundin = paste(getwd(), "tracts", sep = "/"),
  userout = paste(getwd(), "geomasked", sep = "/")
)
confirmGMsettings(maskvars = maskvars, filevars = filevars)

}

}
