% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomask.R
\docType{package}
\name{geomask}
\alias{geomask}
\alias{geomask-package}
\title{geomask: A package for Geostatistical modeling of geomasked data.}
\description{
The geomask package provides three categories of important functions:
foo, bar and baz.
}
\section{Geomaks functions}{

The geomask functions ...
}

