% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGMcompare.R
\name{plotGMcompare}
\alias{plotGMcompare}
\title{Plot GM Comparison Map}
\usage{
plotGMcompare(
  bound,
  original = NULL,
  shifted = NULL,
  maskvars,
  closemap = FALSE
)
}
\arguments{
\item{bound}{Boundary layer.}

\item{original}{Original points or areas spatial layer.}

\item{shifted}{Shifted points or areas spatial layer.}

\item{maskvars}{List of settings for geomasking, including unit type and
minimum, and maximum values.}

\item{closemap}{Boolean denoting whether to close the map window after
the map is drawn and saved.}
}
\description{
This function draws two maps on top of each other with only the polygon
boundaries visible, so that polygon sizes and compositions can be compared.
It is designed to be saved to a recordPlot object for later rendering in a
PDF rather than displayed directly.
}
\details{
If plotting directly, the display may be distorted. For best results, save
to a recordPlot object and write to PDF.
}
