% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGMprogram.R
\name{runGMprogram}
\alias{runGMprogram}
\title{Run the geomasker}
\usage{
runGMprogram(bgcol = "thistle1", buttoncol = "plum2", settings = NULL)
}
\arguments{
\item{bgcol}{Text string containing UI background color.}

\item{buttoncol}{Text string containing UI button color.}

\item{settings}{The filepath to an Rdata file previously created by the
Geomasker.}
}
\description{
Run the geomasker
}
